+++ pom.xml
@@ -60,7 +60,7 @@
                         <!-- Default configuration for running with: mvn clean javafx:run -->
                         <id>default-cli</id>
                         <configuration>
-                            <mainClass>dk.harning.chess_demo/dk.harning.chess_demo.GUI</mainClass>
+                            <mainClass>dk.ek.chess_demo/dk.ek.chess_bot.GUI</mainClass>
                             <launcher>app</launcher>
                             <jlinkZipName>app</jlinkZipName>
                             <jlinkImageName>app</jlinkImageName>
