+++ src/main/java/dk/ek/chess_bot/GUI.java
@@ -1,4 +1,4 @@
-package dk.harning.chess_demo;
+package dk.ek.chess_bot;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
