+++ src/main/java/dk/ek/chess_bot/engine/Board.java
@@ -1,6 +1,6 @@
-package dk.harning.chess_demo.engine;
+package dk.ek.chess_bot.engine;
 
-import static dk.harning.chess_demo.engine.Pieces.*;
+import static dk.ek.chess_bot.engine.Pieces.*;
 
 public class Board {
