+++ src/main/java/dk/ek/chess_bot/engine/Fen.java
@@ -1,8 +1,6 @@
-package dk.harning.chess_demo.engine;
+package dk.ek.chess_bot.engine;
 
-import java.util.Arrays;
-
-import static dk.harning.chess_demo.engine.Pieces.*;
+import static dk.ek.chess_bot.engine.Pieces.*;
 
 public class Fen {
