+++ src/main/java/dk/ek/chess_bot/engine/Pieces.java
@@ -1,4 +1,4 @@
-package dk.harning.chess_demo.engine;
+package dk.ek.chess_bot.engine;
 
 public class Pieces {
     public static final boolean WHITE = true;
