+++ src/main/java/dk/ek/chess_bot/engine/pieces/Knight.java
@@ -1,6 +1,6 @@
-package dk.harning.chess_demo.engine.pieces;
+package dk.ek.chess_bot.engine.pieces;
 
-public class Queen {
+public class Knight {
 
     public static int[] getMoves(){
         return new int[1];
