+++ src/main/java/dk/ek/chess_bot/engine/pieces/Piece.java
@@ -1,4 +1,4 @@
-package dk.harning.chess_demo.engine.pieces;
+package dk.ek.chess_bot.engine.pieces;
 
 public interface Piece {
